requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.5
django-environ>=0.11
django-cors-headers>=4.3
django-storages[boto3]>=1.14
boto3>=1.28
psycopg2-binary>=2.9
Pillow>=10.0
requests>=2.31
